boxinst/hello_world/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.10.2+cpu
torchvision==0.11.3+cpu
opencv-python
boto3
requests
imageio
python-dotenv